+++ circuits/types/field/src/to_lower_bits.rs
@@ -62,7 +62,7 @@ impl<E: Environment> ToLowerBits for Field<E> {
 }
 
 impl<E: Environment> Metrics<dyn ToLowerBits<Boolean = Boolean<E>>> for Field<E> {
-    type Case = (Mode, usize);
+    type Case = (Mode, u64);
 
     fn count(case: &Self::Case) -> Count {
         match case {
