+++ circuits/types/integers/src/add_checked.rs
@@ -82,7 +82,7 @@ impl<E: Environment, I: IntegerType> AddChecked<Self> for Integer<E, I> {
             let sum = self.to_field() + other.to_field();
 
             // Extract the integer bits from the field element, with a carry bit.
-            let (sum, carry) = match sum.to_lower_bits_le(I::BITS + 1).split_last() {
+            let (sum, carry) = match sum.to_lower_bits_le(I::BITS as usize + 1).split_last() {
                 Some((carry, bits_le)) => (Integer::from_bits_le(bits_le), carry.clone()),
                 None => E::halt("Malformed sum detected during integer addition"),
             };
